use execvp instead of subprocess for mysqlimport
If we handoff the execution to the script we are calling instead of
trying to handle it with subprocess we avoid complex situations to pass
thru signals and such

Changed the behaviour, if only one parameter is given we expect that the
sql data is redirected towards the mysql process run inside the
container. When 2 parameters are given the previous behaviour is kept,
where the sql file is copied into the running mysql container and then
sourced inside the container. When all goes well the sql file is removed
from the container to reduce overhead in disk space usage.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>